refactor: share worst-case pre-validation fixtures between bench and timing test
Move insert_decoy_nullifiers and worst_case_bundle into a single
cfg(test)/runtime-benchmarks module so the ignored calibration harness and the
runtime benchmarks cannot drift to different inputs.
