+++ pallets/wormhole/src/tests.rs
@@ -2121,7 +2121,10 @@ mod public_batch_proof_tests {
 /// `cargo test -p pallet-wormhole --release measure_pre_validation_time -- --ignored --nocapture`
 #[cfg(test)]
 mod pre_validation_timing {
-	use crate::mock::*;
+	use crate::{
+		bench_fixtures::{insert_decoy_nullifiers, worst_case_bundle},
+		mock::*,
+	};
 	use qp_wormhole_verifier::{parse_private_batch_public_inputs, ProofWithPublicInputs, C, D, F};
 	use sp_core::H256;
 	use std::time::Instant;
@@ -2132,50 +2135,6 @@ mod pre_validation_timing {
 	fn insert_block_hash(block_number: u32, block_hash: &[u8]) {
 		let bytes: [u8; 32] = block_hash.try_into().unwrap();
 		frame_system::BlockHash::<Test>::insert(block_number as u64, H256::from(bytes));
-	}
-
-	fn insert_decoy_nullifiers(count: u32) {
-		for i in 0..count {
-			let mut nullifier = [0xFFu8; 32];
-			nullifier[0..4].copy_from_slice(&i.to_le_bytes());
-			crate::UsedNullifiers::<Test>::insert(nullifier, true);
-		}
-	}
-
-	fn worst_case_bundle(
-		block_data: qp_wormhole_verifier::BlockData,
-		num_segments: usize,
-	) -> crate::ExitBundle {
-		let slots_per_segment = crate::circuit_config::NUM_LEAF_PROOFS * 2;
-		let mut counter: u32 = 0;
-		let segments = (0..num_segments)
-			.map(|_| {
-				let nullifiers = (0..crate::circuit_config::NUM_LEAF_PROOFS)
-					.map(|_| {
-						counter += 1;
-						let mut bytes = [0xABu8; 32];
-						bytes[0..4].copy_from_slice(&counter.to_le_bytes());
-						qp_wormhole_verifier::BytesDigest::new_unchecked(bytes)
-					})
-					.collect();
-				let account_data = (0..slots_per_segment)
-					.map(|_| qp_wormhole_verifier::PublicInputsByAccount {
-						summed_output_amount: 1,
-						exit_account: qp_wormhole_verifier::BytesDigest::new_unchecked(
-							[0xCDu8; 32],
-						),
-					})
-					.collect();
-				crate::ExitSegment { account_data, nullifiers }
-			})
-			.collect();
-		crate::ExitBundle {
-			asset_id: 0,
-			volume_fee_bps: <Test as crate::Config>::VolumeFeeRateBps::get(),
-			block_data,
-			aggregator_address: None,
-			segments,
-		}
 	}
 
 	#[test]
@@ -2198,7 +2157,7 @@ mod pre_validation_timing {
 				inputs.block_data.block_number,
 				inputs.block_data.block_hash.as_ref(),
 			);
-			insert_decoy_nullifiers(crate::circuit_config::NUM_LEAF_PROOFS as u32);
+			insert_decoy_nullifiers::<Test>(crate::circuit_config::NUM_LEAF_PROOFS as u32);
 
 			let t = Instant::now();
 			for _ in 0..iters {
@@ -2216,7 +2175,7 @@ mod pre_validation_timing {
 			}
 			println!("private full pre-validate: {:?}/iter", t.elapsed() / iters);
 
-			let worst = worst_case_bundle(inputs.block_data.clone(), 1);
+			let worst = worst_case_bundle::<Test>(inputs.block_data.clone(), 1);
 			let t = Instant::now();
 			for _ in 0..iters {
 				assert!(Wormhole::pre_validate_private_batch_proof(&proof_bytes).is_ok());
@@ -2243,7 +2202,7 @@ mod pre_validation_timing {
 				inputs.block_data.block_number,
 				inputs.block_data.block_hash.as_ref(),
 			);
-			insert_decoy_nullifiers(
+			insert_decoy_nullifiers::<Test>(
 				(crate::circuit_config::NUM_PRIVATE_BATCH_PROOFS *
 					crate::circuit_config::NUM_LEAF_PROOFS) as u32,
 			);
@@ -2264,7 +2223,7 @@ mod pre_validation_timing {
 			}
 			println!("public full pre-validate: {:?}/iter", t.elapsed() / iters);
 
-			let worst = worst_case_bundle(
+			let worst = worst_case_bundle::<Test>(
 				inputs.block_data.clone(),
 				crate::circuit_config::NUM_PRIVATE_BATCH_PROOFS,
 			);
